Refact (Dialogs): Combines dialogs into reusable components.
