+++ Breeze.Web/tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/components/auth/authbutton.tsx","./src/components/navigation/navigation.tsx","./src/components/reusable/confirmationmodal.tsx","./src/components/theme/modetoggle.tsx","./src/components/theme/themeprovider.tsx","./src/components/ui/badge.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/carousel.tsx","./src/components/ui/checkbox.tsx","./src/components/ui/dialog.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/form.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/menubar.tsx","./src/components/ui/progress.tsx","./src/components/ui/select.tsx","./src/components/ui/table.tsx","./src/components/ui/tabs.tsx","./src/config/auth/msalinstance.ts","./src/config/environment/useenvironmentvariables.ts","./src/config/routing/approutes.tsx","./src/lib/utils.ts","./src/sections/landingpage.tsx","./src/sections/budget/createbudgetdialog.tsx","./src/sections/budget/editbudgetdialog.tsx","./src/sections/dashboard/dashboard.tsx","./src/sections/dashboard/datatables/expensetable.tsx","./src/sections/dashboard/datatables/incometable.tsx","./src/sections/dashboard/dialogs/expensedialog.tsx","./src/sections/dashboard/dialogs/incomedialog.tsx","./src/sections/goals/goalitem.tsx","./src/sections/goals/goals.tsx","./src/sections/goals/dialogs/creategoaldialog.tsx","./src/sections/goals/dialogs/editgoaldialog.tsx","./src/services/hooks/usehttp.ts","./src/services/hooks/budget/budgetservices.ts","./src/services/hooks/budget/usefetchbudget.ts","./src/services/hooks/category/categoryservices.ts","./src/services/hooks/category/usedeletecategory.ts","./src/services/hooks/category/usefetchcategories.ts","./src/services/hooks/category/usepatchcategory.ts","./src/services/hooks/category/usepostcategory.ts","./src/services/hooks/expense/expenseservices.ts","./src/services/hooks/expense/usedeleteexpense.ts","./src/services/hooks/expense/usefetchexpenses.ts","./src/services/hooks/expense/usepatchexpense.ts","./src/services/hooks/expense/usepostexpense.ts","./src/services/hooks/goal/goalservices.ts","./src/services/hooks/goal/usedeletegoal.ts","./src/services/hooks/goal/usefetchgoals.ts","./src/services/hooks/goal/usepatchgoal.ts","./src/services/hooks/goal/usepostgoal.ts","./src/services/hooks/income/incomeservices.ts","./src/services/hooks/income/usedeleteincome.ts","./src/services/hooks/income/usefetchincomes.ts","./src/services/hooks/income/usepatchincome.ts","./src/services/hooks/income/usepostincome.ts","./src/services/providers/budgetprovider.tsx","./src/services/testdata/testincomes.ts","./src/services/testdata/testbudget.ts","./src/services/testdata/testcategories.ts","./src/services/testdata/testexpenses.ts","./src/services/testdata/testgoals.ts"],"version":"5.8.2"}
+{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/components/auth/authbutton.tsx","./src/components/navigation/navigation.tsx","./src/components/reusable/deleteconfirmationdialog.tsx","./src/components/theme/modetoggle.tsx","./src/components/theme/themeprovider.tsx","./src/components/ui/badge.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/carousel.tsx","./src/components/ui/checkbox.tsx","./src/components/ui/dialog.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/form.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/menubar.tsx","./src/components/ui/progress.tsx","./src/components/ui/select.tsx","./src/components/ui/table.tsx","./src/components/ui/tabs.tsx","./src/config/auth/msalinstance.ts","./src/config/environment/useenvironmentvariables.ts","./src/config/routing/approutes.tsx","./src/lib/utils.ts","./src/sections/landingpage.tsx","./src/sections/budget/budgetdialog.tsx","./src/sections/budget/budgetexpenseitem.tsx","./src/sections/budget/budgetincomeitem.tsx","./src/sections/dashboard/dashboard.tsx","./src/sections/dashboard/datatables/expensetable.tsx","./src/sections/dashboard/datatables/incometable.tsx","./src/sections/dashboard/dialogs/expensedialog.tsx","./src/sections/dashboard/dialogs/incomedialog.tsx","./src/sections/goals/goaldialog.tsx","./src/sections/goals/goals.tsx","./src/services/hooks/usehttp.ts","./src/services/hooks/budget/budgetservices.ts","./src/services/hooks/budget/usefetchbudget.ts","./src/services/hooks/category/categoryservices.ts","./src/services/hooks/category/usedeletecategory.ts","./src/services/hooks/category/usefetchcategories.ts","./src/services/hooks/category/usepatchcategory.ts","./src/services/hooks/category/usepostcategory.ts","./src/services/hooks/expense/expenseservices.ts","./src/services/hooks/expense/usedeleteexpense.ts","./src/services/hooks/expense/usefetchexpensesforbudget.tsx","./src/services/hooks/expense/usefetchexpensesforcategory.ts","./src/services/hooks/expense/usepatchexpense.ts","./src/services/hooks/expense/usepostexpense.ts","./src/services/hooks/goal/goalservices.ts","./src/services/hooks/goal/usedeletegoal.ts","./src/services/hooks/goal/usefetchgoals.ts","./src/services/hooks/goal/usepatchgoal.ts","./src/services/hooks/goal/usepostgoal.ts","./src/services/hooks/income/incomeservices.ts","./src/services/hooks/income/usedeleteincome.ts","./src/services/hooks/income/usefetchincomes.ts","./src/services/hooks/income/usepatchincome.ts","./src/services/hooks/income/usepostincome.ts","./src/services/providers/budgetprovider.tsx","./src/services/testdata/testincomes.ts","./src/services/testdata/testbudget.ts","./src/services/testdata/testcategories.ts","./src/services/testdata/testexpenses.ts","./src/services/testdata/testgoals.ts"],"version":"5.8.2"}
